修正现有代码的编译警告和测试错误
测试错误是
<bean class="org.springframework.session.data.redis.config.annotation.web.http.RedisHttpSessionConfiguration"/>
导致的，springSessionRepositoryFilter 是在web.xml中配置的，测试时找不到实例，现在测试配置文件中将该代码注释，等待进一步修正配置文件。

+++ spark-core/spark-common/src/main/java/net/parim/common/persistence/mybatis/paginate/OffsetLimitInterceptor.java
@@ -122,7 +122,7 @@ public class OffsetLimitInterceptor implements Interceptor {
 			invocation.getArgs()[0] = newMs;
 			//org.springframework.data.domain.Page page2;
 			@SuppressWarnings("unchecked")
-			PageImpl<?> pageImpl = new PageImpl<Object>((List<Object>)invocation.proceed(), pageable, (long)page.getCount());
+			PageImpl<?> pageImpl = new PageImpl<Object>((List<Object>)invocation.proceed(), pageable, page.getCount());
 			return pageImpl;
 		}
